topic_zoom: Merge into stream_list.
