+++ tools/test-js-with-node
@@ -233,7 +233,6 @@ EXEMPT_FILES = make_set(
         "web/src/todo_widget.js",
         "web/src/topic_list.js",
         "web/src/topic_popover.js",
-        "web/src/topic_zoom.js",
         "web/src/tutorial.js",
         "web/src/types.ts",
         "web/src/typing.js",
