+++ web/src/ui_init.js
@@ -124,7 +124,6 @@ import * as timerender from "./timerender";
 import * as tippyjs from "./tippyjs";
 import * as topic_list from "./topic_list";
 import * as topic_popover from "./topic_popover";
-import * as topic_zoom from "./topic_zoom";
 import * as transmit from "./transmit";
 import * as tutorial from "./tutorial";
 import * as typeahead_helper from "./typeahead_helper";
@@ -713,7 +712,6 @@ export function initialize_everything() {
             );
         },
     });
-    topic_zoom.initialize();
     drafts.initialize();
     drafts_overlay_ui.initialize();
     sent_messages.initialize();
